Move extension methods closer to serialisation classes
