+++ osu.Game/Skinning/ISerialisableDrawableContainer.cs
@@ -5,7 +5,6 @@ using System.Collections.Generic;
 using System.Linq;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
-using osu.Game.Extensions;
 
 namespace osu.Game.Skinning
 {
